requirements.txt
openai
python-dotenv
requests
flask
python-dotenv